Projeto Final

+++ projeto/script.js
@@ -2,9 +2,19 @@ import TodoModel from "./TodoModel.js";
 
 window.onload = async () => {
 
-    let currentTaskIndex;
-
+    //SERVICE WORKER
+    if ("serviceWorker" in navigator) {
+        try {
+            await navigator.serviceWorker.register("serviceWorker.js");
+        } catch (error) {
+            console.error("Error registering Service Worker:", error);
+        }
+    } else {
+        alert("service worker not supported, use a chromium based browser.");
+    }
+    
     const model = new TodoModel();
+    let currentTaskIndex;
 
     // O teu código aqui...
     const listsContainer = document.querySelector("#lists-container");
@@ -12,58 +22,90 @@ window.onload = async () => {
     todoHeader.addEventListener("clicked", () => {
         listsContainer.style.transform = "translateX(0)";
         todoHeader.state = "tasks";
-        buildTasksList(model.getTasks());
     });
+    todoHeader.state = "tasks";
 
-    
+    const todoDialog = document.querySelector("todo-dialog");
+    todoDialog.addEventListener("confirm", (ev) => {
+        if(todoHeader.state === "tasks") {
+            model.addTask({
+                title: ev.detail.value,
+                items: []
+            });
+            buildTasksList(model.getTasks());
+        } else {
+            model.addItem(currentTaskIndex, {
+                title: ev.detail.value,
+                checked: "false"
+            });
+
+            buildItemsList(currentTaskIndex, model.getItems(currentTaskIndex));
+        }
+    })
+    const footer = document.querySelector("footer");
+    footer.onclick = () => {
+        const val = todoHeader.state === "tasks" ? "Add Task" : "Add Item";
+        todoDialog.open(val);
+    };
+
     const buildTasksList = (tasks) => {
-        const tasksList = document.querySelector("#tasks");
-        tasksList.innerHTML = "";
+        const ul = document.querySelector("#tasks");
+        ul.innerHTML = "";
 
+        if(tasks.length === 0) {
+            ul.innerHTML = "<li class='empty'>No tasks</li>";
+            return;
+        }
         tasks.forEach((task, index) => {
             const li = document.createElement("li");
             const taskItem = new TaskItem();
+            taskItem.title = task.title;
             taskItem.addEventListener("clicked", () => {
                 listsContainer.style.transform = "translateX(-100%)";
                 todoHeader.state = "items";
-                todoHeader.taskName = task.title;
-                buildItemsList(task.items);
+                todoHeader.taskName = taskItem.title;
+                const itemsData = model.getItems(index);
+                buildItemsList(index, itemsData);
                 currentTaskIndex = index;
             });
             taskItem.addEventListener("delete", () => {
                 model.deleteTask(index);
                 buildTasksList(model.getTasks());
             });
-            taskItem.title = task.title;
-
             li.append(taskItem);
-            tasksList.append(li);
+            ul.append(li);
         });
-
     }
+    const buildItemsList = (taskID, items) => {
 
-    const buildItemsList = (items) => {
-
-        const checkItemsList = document.querySelector("#items");
-        checkItemsList.innerHTML = "";
+        const ul = document.querySelector("#items");
+        ul.innerHTML = "";
+        if(items.length === 0) {
+            ul.innerHTML = "<li class='empty'>No items</li>";
+            return;
+        }
         items.forEach((item, index) => {
             const li = document.createElement("li");
             const checkItem = new CheckItem();
-            checkItem.addEventListener("checked", (ev) => {
-                model.updateItem(currentTaskIndex, index, ev.detail.checked);
-            });
-            checkItem.addEventListener("delete", () => {
-                model.deleteItem(currentTaskIndex, index);
-                buildItemsList(model.getItems(currentTaskIndex));
-            })
+            
             checkItem.title = item.title;
             checkItem.checked = item.checked;
+            checkItem.addEventListener("checked", (ev) => {
+                model.updateItem(taskID, index, ev.detail.checked);
+                buildTasksList(model.getTasks());
+            });
+            checkItem.addEventListener("delete", () => {
 
+                model.deleteItem(currentTaskIndex, index);
+                buildItemsList(currentTaskIndex, model.getItems(currentTaskIndex));
+                console.log("delete item");
+            });
             li.append(checkItem);
-            checkItemsList.append(li);
-            
+            ul.append(li);
         });
     }
-
     buildTasksList(model.getTasks());
+
+
+    document.body.style.opacity = 1;
 }
